refactor: moved chan box move logic to new hook
